feat(FSADT1-1365): new endpoints added
