+++ legacy/src/main/java/ca/bc/gov/app/dto/AddressSearchDto.java
@@ -1,0 +1,17 @@
+package ca.bc.gov.app.dto;
+
+import org.apache.commons.lang3.StringUtils;
+
+public record AddressSearchDto(
+    String address,
+    String city,
+    String province,
+    String postalCode,
+    String country
+) {
+
+  public boolean isValid() {
+    return !StringUtils.isAnyBlank(address, city, province, postalCode, country);
+  }
+
+}
